sens_pozitiv/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
arrow==1.2.2
    # via django-q
asgiref==3.5.0
    # via django
blessed==1.19.1
    # via django-q
brotli==1.0.9
    # via flask-compress
certifi==2021.10.8
    # via requests
charset-normalizer==2.0.12
    # via requests
click==8.0.4
    # via flask
dash==1.20.0
    # via
    #   dash-bootstrap-components
    #   django-plotly-dash
dash-bootstrap-components==0.13.0
    # via -r requirements.in
dash-core-components==1.16.0
    # via
    #   dash
    #   django-plotly-dash
dash-html-components==1.1.3
    # via
    #   dash
    #   django-plotly-dash
dash-renderer==1.9.1
    # via
    #   dash
    #   django-plotly-dash
dash-table==4.11.3
    # via dash
defusedxml==0.7.1
    # via odfpy
diff-match-patch==20200713
    # via django-import-export
dj-database-url==0.5.0
    # via -r requirements.in
django==3.2.12
    # via
    #   -r requirements.in
    #   django-import-export
    #   django-jazzmin
    #   django-js-asset
    #   django-multiselectfield
    #   django-phonenumber-field
    #   django-picklefield
    #   django-plotly-dash
    #   django-q
    #   django-recaptcha
    #   django-storages
django-ckeditor==6.2.0
    # via -r requirements.in
django-crispy-forms==1.14.0
    # via -r requirements.in
django-environ==0.8.1
    # via -r requirements.in
django-import-export==2.7.1
    # via -r requirements.in
django-jazzmin==2.4.9
    # via -r requirements.in
django-js-asset==2.0.0
    # via django-ckeditor
django-multiselectfield==0.1.12
    # via -r requirements.in
django-phonenumber-field[phonenumbers]==6.1.0
    # via -r requirements.in
django-picklefield==3.0.1
    # via django-q
django-plotly-dash==1.6.6
    # via -r requirements.in
django-q==1.3.9
    # via
    #   -r requirements.in
    #   django-q-email
django-q-email==5.0.1
    # via -r requirements.in
django-recaptcha==3.0.0
    # via -r requirements.in
django-storages==1.12.3
    # via -r requirements.in
dpd-components==0.1.0
    # via
    #   -r requirements.in
    #   django-plotly-dash
dpd-static-support==0.0.5
    # via -r requirements.in
et-xmlfile==1.1.0
    # via openpyxl
flask==2.0.3
    # via
    #   dash
    #   django-plotly-dash
    #   flask-compress
flask-compress==1.11
    # via dash
future==0.18.2
    # via dash
gevent==21.12.0
    # via -r requirements.in
greenlet==1.1.2
    # via gevent
gunicorn==20.1.0
    # via -r requirements.in
idna==3.3
    # via requests
itsdangerous==2.1.1
    # via flask
jinja2==3.0.3
    # via flask
markuppy==1.14
    # via tablib
markupsafe==2.1.1
    # via jinja2
numpy==1.22.3
    # via pandas
odfpy==1.4.1
    # via tablib
openpyxl==3.0.9
    # via tablib
pandas==1.4.1
    # via -r requirements.in
phonenumbers==8.12.45
    # via django-phonenumber-field
plotly==5.6.0
    # via
    #   dash
    #   django-plotly-dash
psycopg2==2.9.3
    # via -r requirements.in
python-dateutil==2.8.2
    # via
    #   arrow
    #   pandas
pytz==2021.3
    # via
    #   django
    #   pandas
pyyaml==6.0
    # via tablib
redis==3.5.3
    # via django-q
requests==2.27.1
    # via -r requirements.in
six==1.16.0
    # via
    #   blessed
    #   plotly
    #   python-dateutil
sqlparse==0.4.2
    # via django
supervisor==4.2.4
    # via -r requirements.in
tablib[html,ods,xls,xlsx,yaml]==3.2.0
    # via django-import-export
tenacity==8.0.1
    # via plotly
urllib3==1.26.8
    # via requests
wcwidth==0.2.5
    # via blessed
werkzeug==2.0.3
    # via flask
whitenoise==6.0.0
    # via -r requirements.in
xlrd==2.0.1
    # via tablib
xlsxwriter==3.0.3
    # via -r requirements.in
xlwt==1.3.0
    # via tablib
zope-event==4.5.0
    # via gevent
zope-interface==5.4.0
    # via gevent

# The following packages are considered to be unsafe in a requirements file:
# setuptools
